requirements.txt
pip-chill==1.0.1
torchaudio==0.12.1
torchvision==0.2.2
